+++ app/src/main/java/com/dev/briefing/data/datasource/BriefingDataSource.kt
@@ -14,5 +14,5 @@ interface BriefingDataSource {
         timeOfDay: TimeOfDay? = null
     ): BriefingCategoryArticlesResponse
 
-    suspend fun getBriefingArticle(id: Int): BriefingArticleResponse
+    suspend fun getBriefingArticle(id: Long): BriefingArticleResponse
 }
