+++ app/src/main/java/com/dev/briefing/data/respository/BriefingRepository.kt
@@ -22,5 +22,5 @@ interface BriefingRepository {
         timeOfDay: TimeOfDay? = null
     ): BriefingCategoryArticles
 
-    suspend fun getBriefingDetail(id: Int): BriefingArticle
+    suspend fun getBriefingDetail(id: Long): BriefingArticle
 }
